+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 ### Added
 - Authentication - Users can connect to their project using Google Single Sign-On.
 
+### Changed
+- Dependencies - Upgrade the json gem to work with Ruby versions 2.4.X on OSX.
+
 ## RELEASE 2.0.4 - 2018-01-08
 ### Changed
 - Performance - Set the CORS Max-Age to 1 day to be consistent with the other lianas behaviour.
-- Dependencies - Upgrade the json gem to work with Ruby versions 2.4.X on OSX.
 
 ### Fixed
 - Serializer - Fix 'already defined' warning message from serializer on server start.